+++ tsl/src/compression/compression.h
@@ -153,4 +153,9 @@ extern DecompressionIterator *(*tsl_get_decompression_iterator_init(
 extern void update_compressed_chunk_relstats(Oid uncompressed_relid, Oid compressed_relid);
 extern void merge_chunk_relstats(Oid merged_relid, Oid compressed_relid);
 
+typedef struct Chunk Chunk;
+typedef struct ChunkInsertState ChunkInsertState;
+extern void decompress_batches_for_insert(ChunkInsertState *cis, Chunk *chunk,
+										  TupleTableSlot *slot);
+
 #endif